Crate async_rx

Utility functions for async reactive programming.

This crate is intentionally very small as it only provides utilities that are not already found in futures-util. It is meant as a supplement, not a replacement for the existing well-known futures crates.

Structs

BatchWith

Stream adapter produced by StreamExt::batch_with.

Dedup

Stream adapter produced by StreamExt::dedup.

DedupByKey

Stream adapter produced by StreamExt::dedup_by_key.

Switch

Stream adapter produced by StreamExt::switch.

Traits

StreamExt

Extensions to the Stream trait.